requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.1.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0
python-dotenv>=0.19.0
pathlib>=1.0.1
mlflow>=1.20.0
fastapi>=0.68.0
uvicorn>=0.15.0
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.0
pre-commit>=2.15.0
pyyaml>=5.4.1
python-multipart>=0.0.5
fpdf>=1.7.2
jinja2>=3.0.0
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 